+++ src/aaccoder/trellis.rs
@@ -39,12 +39,8 @@ pub(crate) unsafe fn codebook_rate(
     let mut i: c_int = 0;
     let mut j: c_int = 0;
     let max_sfb: c_int = (*sce).ics.max_sfb as c_int;
-    let run_bits: c_int = if (*sce).ics.num_windows == 1 as c_int {
-        5 as c_int
-    } else {
-        3 as c_int
-    };
-    let run_esc: c_int = ((1 as c_int) << run_bits) - 1 as c_int;
+    let run_bits: c_int = if (*sce).ics.num_windows == 1 { 5 } else { 3 };
+    let run_esc: c_int = ((1) << run_bits) - 1;
     let mut idx: c_int = 0;
     let mut ppos: c_int = 0;
     let mut count: c_int = 0;
@@ -52,99 +48,95 @@ pub(crate) unsafe fn codebook_rate(
     let mut stackcb: [c_int; 120] = [0; 120];
     let mut stack_len: c_int = 0;
     let mut next_minbits: c_float = ::core::f32::INFINITY;
-    let mut next_mincb: c_int = 0 as c_int;
+    let mut next_mincb: c_int = 0;
     abs_pow34_v(
         ((*s).scoefs).as_mut_ptr(),
         ((*sce).coeffs).as_mut_ptr(),
-        1024 as c_int,
+        1024,
     );
-    start = win * 128 as c_int;
-    cb = 0 as c_int;
-    while cb < 15 as c_int {
-        path[0 as c_int as usize][cb as usize].cost = (run_bits + 4 as c_int) as c_float;
-        path[0 as c_int as usize][cb as usize].prev_idx = -(1 as c_int);
-        path[0 as c_int as usize][cb as usize].run = 0 as c_int;
+    start = win * 128;
+    cb = 0;
+    while cb < 15 {
+        path[0][cb as usize].cost = (run_bits + 4) as c_float;
+        path[0][cb as usize].prev_idx = -1;
+        path[0][cb as usize].run = 0;
         cb += 1;
         cb;
     }
-    swb = 0 as c_int;
+    swb = 0;
     while swb < max_sfb {
         size = *((*sce).ics.swb_sizes).offset(swb as isize) as c_int;
-        if (*sce).zeroes[(win * 16 as c_int + swb) as usize] != 0 {
-            let mut cost_stay_here: c_float = path[swb as usize][0 as c_int as usize].cost;
-            let mut cost_get_here: c_float =
-                next_minbits + run_bits as c_float + 4 as c_int as c_float;
-            if *(run_value_bits[((*sce).ics.num_windows == 8 as c_int) as c_int as usize])
-                .offset(path[swb as usize][0 as c_int as usize].run as isize)
-                as c_int
-                != *(run_value_bits[((*sce).ics.num_windows == 8 as c_int) as c_int as usize])
-                    .offset((path[swb as usize][0 as c_int as usize].run + 1 as c_int) as isize)
-                    as c_int
+        if (*sce).zeroes[(win * 16 + swb) as usize] != 0 {
+            let mut cost_stay_here: c_float = path[swb as usize][0].cost;
+            let mut cost_get_here: c_float = next_minbits + run_bits as c_float + 4 as c_float;
+            if *(run_value_bits[((*sce).ics.num_windows == 8) as c_int as usize])
+                .offset(path[swb as usize][0].run as isize) as c_int
+                != *(run_value_bits[((*sce).ics.num_windows == 8) as c_int as usize])
+                    .offset((path[swb as usize][0].run + 1) as isize) as c_int
             {
                 cost_stay_here += run_bits as c_float;
             }
             if cost_get_here < cost_stay_here {
-                path[(swb + 1 as c_int) as usize][0 as c_int as usize].prev_idx = next_mincb;
-                path[(swb + 1 as c_int) as usize][0 as c_int as usize].cost = cost_get_here;
-                path[(swb + 1 as c_int) as usize][0 as c_int as usize].run = 1 as c_int;
+                path[(swb + 1) as usize][0].prev_idx = next_mincb;
+                path[(swb + 1) as usize][0].cost = cost_get_here;
+                path[(swb + 1) as usize][0].run = 1;
             } else {
-                path[(swb + 1 as c_int) as usize][0 as c_int as usize].prev_idx = 0 as c_int;
-                path[(swb + 1 as c_int) as usize][0 as c_int as usize].cost = cost_stay_here;
-                path[(swb + 1 as c_int) as usize][0 as c_int as usize].run =
-                    path[swb as usize][0 as c_int as usize].run + 1 as c_int;
+                path[(swb + 1) as usize][0].prev_idx = 0;
+                path[(swb + 1) as usize][0].cost = cost_stay_here;
+                path[(swb + 1) as usize][0].run = path[swb as usize][0].run + 1;
             }
-            next_minbits = path[(swb + 1 as c_int) as usize][0 as c_int as usize].cost;
-            next_mincb = 0 as c_int;
-            cb = 1 as c_int;
-            while cb < 15 as c_int {
-                path[(swb + 1 as c_int) as usize][cb as usize].cost = 61450 as c_int as c_float;
-                path[(swb + 1 as c_int) as usize][cb as usize].prev_idx = -(1 as c_int);
-                path[(swb + 1 as c_int) as usize][cb as usize].run = 0 as c_int;
+            next_minbits = path[(swb + 1) as usize][0].cost;
+            next_mincb = 0;
+            cb = 1;
+            while cb < 15 {
+                path[(swb + 1) as usize][cb as usize].cost = 61450 as c_float;
+                path[(swb + 1) as usize][cb as usize].prev_idx = -1;
+                path[(swb + 1) as usize][cb as usize].run = 0;
                 cb += 1;
                 cb;
             }
         } else {
             let mut minbits: c_float = next_minbits;
             let mut mincb: c_int = next_mincb;
-            let mut startcb: c_int = (*sce).band_type[(win * 16 as c_int + swb) as usize] as c_int;
+            let mut startcb: c_int = (*sce).band_type[(win * 16 + swb) as usize] as c_int;
             startcb = aac_cb_in_map[startcb as usize] as c_int;
             next_minbits = ::core::f32::INFINITY;
-            next_mincb = 0 as c_int;
-            cb = 0 as c_int;
+            next_mincb = 0;
+            cb = 0;
             while cb < startcb {
-                path[(swb + 1 as c_int) as usize][cb as usize].cost = 61450 as c_int as c_float;
-                path[(swb + 1 as c_int) as usize][cb as usize].prev_idx = -(1 as c_int);
-                path[(swb + 1 as c_int) as usize][cb as usize].run = 0 as c_int;
+                path[(swb + 1) as usize][cb as usize].cost = 61450 as c_float;
+                path[(swb + 1) as usize][cb as usize].prev_idx = -1;
+                path[(swb + 1) as usize][cb as usize].run = 0;
                 cb += 1;
                 cb;
             }
             cb = startcb;
-            while cb < 15 as c_int {
+            while cb < 15 {
                 let mut cost_stay_here_0: c_float = 0.;
                 let mut cost_get_here_0: c_float = 0.;
                 let mut bits: c_float = 0.0f32;
-                if cb >= 12 as c_int
-                    && (*sce).band_type[(win * 16 as c_int + swb) as usize] as c_uint
+                if cb >= 12
+                    && (*sce).band_type[(win * 16 + swb) as usize] as c_uint
                         != aac_cb_out_map[cb as usize] as c_uint
                 {
-                    path[(swb + 1 as c_int) as usize][cb as usize].cost = 61450 as c_int as c_float;
-                    path[(swb + 1 as c_int) as usize][cb as usize].prev_idx = -(1 as c_int);
-                    path[(swb + 1 as c_int) as usize][cb as usize].run = 0 as c_int;
+                    path[(swb + 1) as usize][cb as usize].cost = 61450 as c_float;
+                    path[(swb + 1) as usize][cb as usize].prev_idx = -1;
+                    path[(swb + 1) as usize][cb as usize].run = 0;
                 } else {
-                    w = 0 as c_int;
+                    w = 0;
                     while w < group_len {
                         bits += quantize_band_cost_bits(
                             s,
                             &mut *((*sce).coeffs)
                                 .as_mut_ptr()
-                                .offset((start + w * 128 as c_int) as isize),
+                                .offset((start + w * 128) as isize),
                             &mut *((*s).scoefs)
                                 .as_mut_ptr()
-                                .offset((start + w * 128 as c_int) as isize),
+                                .offset((start + w * 128) as isize),
                             size,
-                            (*sce).sf_idx[(win * 16 as c_int + swb) as usize],
+                            (*sce).sf_idx[(win * 16 + swb) as usize],
                             aac_cb_out_map[cb as usize] as c_int,
-                            0 as c_int as c_float,
+                            0 as c_float,
                             ::core::f32::INFINITY,
                             ptr::null_mut::<c_int>(),
                             ptr::null_mut::<c_float>(),
@@ -153,29 +145,28 @@ pub(crate) unsafe fn codebook_rate(
                         w;
                     }
                     cost_stay_here_0 = path[swb as usize][cb as usize].cost + bits;
-                    cost_get_here_0 = minbits + bits + run_bits as c_float + 4 as c_int as c_float;
-                    if *(run_value_bits[((*sce).ics.num_windows == 8 as c_int) as c_int as usize])
+                    cost_get_here_0 = minbits + bits + run_bits as c_float + 4 as c_float;
+                    if *(run_value_bits[((*sce).ics.num_windows == 8) as c_int as usize])
                         .offset(path[swb as usize][cb as usize].run as isize)
                         as c_int
-                        != *(run_value_bits
-                            [((*sce).ics.num_windows == 8 as c_int) as c_int as usize])
-                            .offset((path[swb as usize][cb as usize].run + 1 as c_int) as isize)
+                        != *(run_value_bits[((*sce).ics.num_windows == 8) as c_int as usize])
+                            .offset((path[swb as usize][cb as usize].run + 1) as isize)
                             as c_int
                     {
                         cost_stay_here_0 += run_bits as c_float;
                     }
                     if cost_get_here_0 < cost_stay_here_0 {
-                        path[(swb + 1 as c_int) as usize][cb as usize].prev_idx = mincb;
-                        path[(swb + 1 as c_int) as usize][cb as usize].cost = cost_get_here_0;
-                        path[(swb + 1 as c_int) as usize][cb as usize].run = 1 as c_int;
+                        path[(swb + 1) as usize][cb as usize].prev_idx = mincb;
+                        path[(swb + 1) as usize][cb as usize].cost = cost_get_here_0;
+                        path[(swb + 1) as usize][cb as usize].run = 1;
                     } else {
-                        path[(swb + 1 as c_int) as usize][cb as usize].prev_idx = cb;
-                        path[(swb + 1 as c_int) as usize][cb as usize].cost = cost_stay_here_0;
-                        path[(swb + 1 as c_int) as usize][cb as usize].run =
-                            path[swb as usize][cb as usize].run + 1 as c_int;
+                        path[(swb + 1) as usize][cb as usize].prev_idx = cb;
+                        path[(swb + 1) as usize][cb as usize].cost = cost_stay_here_0;
+                        path[(swb + 1) as usize][cb as usize].run =
+                            path[swb as usize][cb as usize].run + 1;
                     }
-                    if path[(swb + 1 as c_int) as usize][cb as usize].cost < next_minbits {
-                        next_minbits = path[(swb + 1 as c_int) as usize][cb as usize].cost;
+                    if path[(swb + 1) as usize][cb as usize].cost < next_minbits {
+                        next_minbits = path[(swb + 1) as usize][cb as usize].cost;
                         next_mincb = cb;
                     }
                 }
@@ -187,10 +178,10 @@ pub(crate) unsafe fn codebook_rate(
         swb += 1;
         swb;
     }
-    stack_len = 0 as c_int;
-    idx = 0 as c_int;
-    cb = 1 as c_int;
-    while cb < 15 as c_int {
+    stack_len = 0;
+    idx = 0;
+    cb = 1;
+    while cb < 15 {
         if path[max_sfb as usize][cb as usize].cost < path[max_sfb as usize][idx as usize].cost {
             idx = cb;
         }
@@ -198,35 +189,34 @@ pub(crate) unsafe fn codebook_rate(
         cb;
     }
     ppos = max_sfb;
-    while ppos > 0 as c_int {
+    while ppos > 0 {
         cb = idx;
         stackrun[stack_len as usize] = path[ppos as usize][cb as usize].run;
         stackcb[stack_len as usize] = cb;
-        idx = path[(ppos - path[ppos as usize][cb as usize].run + 1 as c_int) as usize]
-            [cb as usize]
-            .prev_idx;
+        idx =
+            path[(ppos - path[ppos as usize][cb as usize].run + 1) as usize][cb as usize].prev_idx;
         ppos -= path[ppos as usize][cb as usize].run;
         stack_len += 1;
         stack_len;
     }
-    start = 0 as c_int;
-    i = stack_len - 1 as c_int;
-    while i >= 0 as c_int {
+    start = 0;
+    i = stack_len - 1;
+    while i >= 0 {
         cb = aac_cb_out_map[stackcb[i as usize] as usize] as c_int;
-        put_bits(&mut (*s).pb, 4 as c_int, cb as BitBuf);
+        put_bits(&mut (*s).pb, 4, cb as BitBuf);
         count = stackrun[i as usize];
         ptr::write_bytes(
             (*sce)
                 .zeroes
                 .as_mut_ptr()
-                .offset((win * 16 as c_int) as isize)
+                .offset((win * 16) as isize)
                 .offset(start as isize),
             u8::from(cb == 0),
             count as usize,
         );
-        j = 0 as c_int;
+        j = 0;
         while j < count {
-            (*sce).band_type[(win * 16 as c_int + start) as usize] = cb as BandType;
+            (*sce).band_type[(win * 16 + start) as usize] = cb as BandType;
             start += 1;
             start;
             j += 1;
