+++ src/aacenc/window.rs
@@ -25,12 +25,12 @@ unsafe fn apply_only_long_window(
         SINE_WIN_1024.as_ptr()
     };
     let mut out: *mut c_float = ((*sce).ret_buf).as_mut_ptr();
-    ((*fdsp).vector_fmul).expect("non-null function pointer")(out, audio, lwindow, 1024 as c_int);
+    ((*fdsp).vector_fmul).expect("non-null function pointer")(out, audio, lwindow, 1024);
     ((*fdsp).vector_fmul_reverse).expect("non-null function pointer")(
-        out.offset(1024 as c_int as isize),
-        audio.offset(1024 as c_int as isize),
+        out.offset(1024),
+        audio.offset(1024),
         pwindow,
-        1024 as c_int,
+        1024,
     );
 }
 unsafe fn apply_long_start_window(
@@ -49,32 +49,15 @@ unsafe fn apply_long_start_window(
         &SINE_WIN_128
     };
     let mut out: *mut c_float = ((*sce).ret_buf).as_mut_ptr();
-    ((*fdsp).vector_fmul).expect("non-null function pointer")(
-        out,
-        audio,
-        lwindow.as_ptr(),
-        1024 as c_int,
-    );
-    ptr::copy_nonoverlapping(
-        audio.offset(1024 as c_int as isize),
-        out.offset(1024 as c_int as isize),
-        448,
-    );
+    ((*fdsp).vector_fmul).expect("non-null function pointer")(out, audio, lwindow.as_ptr(), 1024);
+    ptr::copy_nonoverlapping(audio.offset(1024), out.offset(1024), 448);
     ((*fdsp).vector_fmul_reverse).expect("non-null function pointer")(
-        out.offset(1024 as c_int as isize)
-            .offset(448 as c_int as isize),
-        audio
-            .offset(1024 as c_int as isize)
-            .offset(448 as c_int as isize),
+        out.offset(1024).offset(448),
+        audio.offset(1024).offset(448),
         swindow.as_ptr(),
-        128 as c_int,
+        128,
     );
-    ptr::write_bytes(
-        out.offset(1024 as c_int as isize)
-            .offset(576 as c_int as isize),
-        0,
-        448,
-    );
+    ptr::write_bytes(out.offset(1024).offset(576), 0, 448);
 }
 unsafe fn apply_long_stop_window(
     mut fdsp: *mut AVFloatDSPContext,
@@ -94,21 +77,17 @@ unsafe fn apply_long_stop_window(
     let mut out: *mut c_float = ((*sce).ret_buf).as_mut_ptr();
     ptr::write_bytes(out, 0, 448);
     ((*fdsp).vector_fmul).expect("non-null function pointer")(
-        out.offset(448 as c_int as isize),
-        audio.offset(448 as c_int as isize),
+        out.offset(448),
+        audio.offset(448),
         swindow,
-        128 as c_int,
+        128,
     );
-    ptr::copy_nonoverlapping(
-        audio.offset(576 as c_int as isize),
-        out.offset(576 as c_int as isize),
-        448,
-    );
+    ptr::copy_nonoverlapping(audio.offset(576), out.offset(576), 448);
     ((*fdsp).vector_fmul_reverse).expect("non-null function pointer")(
-        out.offset(1024 as c_int as isize),
-        audio.offset(1024 as c_int as isize),
+        out.offset(1024),
+        audio.offset(1024),
         lwindow,
-        1024 as c_int,
+        1024,
     );
 }
 unsafe fn apply_eight_short_window(
@@ -126,26 +105,21 @@ unsafe fn apply_eight_short_window(
     } else {
         SINE_WIN_128.as_ptr()
     };
-    let mut in_0: *const c_float = audio.offset(448 as c_int as isize);
+    let mut in_0: *const c_float = audio.offset(448);
     let mut out: *mut c_float = ((*sce).ret_buf).as_mut_ptr();
     let mut w: c_int = 0;
-    w = 0 as c_int;
-    while w < 8 as c_int {
+    w = 0;
+    while w < 8 {
         ((*fdsp).vector_fmul).expect("non-null function pointer")(
             out,
             in_0,
             if w != 0 { pwindow } else { swindow },
-            128 as c_int,
+            128,
         );
-        out = out.offset(128 as c_int as isize);
-        in_0 = in_0.offset(128 as c_int as isize);
-        ((*fdsp).vector_fmul_reverse).expect("non-null function pointer")(
-            out,
-            in_0,
-            swindow,
-            128 as c_int,
-        );
-        out = out.offset(128 as c_int as isize);
+        out = out.offset(128);
+        in_0 = in_0.offset(128);
+        ((*fdsp).vector_fmul_reverse).expect("non-null function pointer")(out, in_0, swindow, 128);
+        out = out.offset(128);
         w += 1;
         w;
     }
@@ -178,18 +152,18 @@ pub(super) unsafe fn apply_window_and_mdct(
             size_of::<c_float>() as c_ulong as ptrdiff_t,
         );
     } else {
-        i = 0 as c_int;
-        while i < 1024 as c_int {
+        i = 0;
+        while i < 1024 {
             ((*s).mdct128_fn).expect("non-null function pointer")(
                 (*s).mdct128,
                 &mut *((*sce).coeffs).as_mut_ptr().offset(i as isize) as *mut c_float
                     as *mut c_void,
-                output.offset((i * 2 as c_int) as isize) as *mut c_void,
+                output.offset((i * 2) as isize) as *mut c_void,
                 size_of::<c_float>() as c_ulong as ptrdiff_t,
             );
-            i += 128 as c_int;
+            i += 128;
         }
     }
-    ptr::copy_nonoverlapping(audio.offset(1024 as c_int as isize), audio, 1024);
+    ptr::copy_nonoverlapping(audio.offset(1024), audio, 1024);
     (*sce).pcoeffs = (*sce).coeffs;
 }
